+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.15 ##
+* Support for sa.BINARY and sa.BLOB types
+* Support nullable StructType fields via Optional wrapper
+
 ## 0.1.14 ##
 * Add ability to propagate RetrySettings